Split tracking service into separate service for ingest only
