+++ src/main/java/com/rackspace/ceres/app/services/DownsampleTrackingService.java
@@ -16,22 +16,11 @@
 
 package com.rackspace.ceres.app.services;
 
-import com.github.benmanes.caffeine.cache.AsyncCache;
-import com.google.common.hash.HashCode;
-import com.rackspace.ceres.app.config.AppProperties;
 import com.rackspace.ceres.app.config.DownsampleProperties;
-import com.rackspace.ceres.app.downsample.TemporalNormalizer;
-import com.rackspace.ceres.app.entities.SeriesSet;
-import com.rackspace.ceres.app.model.DownsampleSetCacheKey;
-import com.rackspace.ceres.app.model.Downsampling;
 import com.rackspace.ceres.app.model.PendingDownsampleSet;
-import com.rackspace.ceres.app.model.TimeslotCacheKey;
 import com.rackspace.ceres.app.utils.DateTimeUtils;
 import lombok.extern.slf4j.Slf4j;
-import org.reactivestreams.Publisher;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.data.cassandra.core.ReactiveCassandraTemplate;
 import org.springframework.data.cassandra.core.cql.ReactiveCqlTemplate;
 import org.springframework.data.redis.core.ReactiveStringRedisTemplate;
 import org.springframework.data.redis.core.script.RedisScript;
@@ -42,7 +31,6 @@ import reactor.core.publisher.Mono;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.List;
-import java.util.concurrent.CompletableFuture;
 
 import static com.rackspace.ceres.app.utils.DateTimeUtils.nowEpochSeconds;
 
@@ -52,37 +40,21 @@ import static com.rackspace.ceres.app.utils.DateTimeUtils.nowEpochSeconds;
 public class DownsampleTrackingService {
 
   private final DownsampleProperties properties;
-  private final SeriesSetService hashService;
   private final ReactiveCqlTemplate cqlTemplate;
-  private final ReactiveCassandraTemplate cassandraTemplate;
   private final ReactiveStringRedisTemplate redisTemplate;
   private final RedisScript<String> redisGetJob;
-  private final AsyncCache<DownsampleSetCacheKey, Boolean> downsampleHashExistenceCache;
-  private final AsyncCache<TimeslotCacheKey, Boolean> timeslotExistenceCache;
 
   private static final String GET_HASHES_TO_DOWNSAMPLE = "SELECT hash FROM downsampling_hashes WHERE partition = ?";
 
-  private final AppProperties appProperties;
-
   @Autowired
   public DownsampleTrackingService(ReactiveStringRedisTemplate redisTemplate,
-                                   ReactiveCassandraTemplate cassandraTemplate,
                                    RedisScript<String> redisGetJob,
                                    DownsampleProperties properties,
-                                   SeriesSetService hashService,
-                                   AppProperties appProperties,
-                                   AsyncCache<DownsampleSetCacheKey, Boolean> downsampleHashExistenceCache,
-                                   AsyncCache<TimeslotCacheKey, Boolean> timeslotExistenceCache,
                                    ReactiveCqlTemplate cqlTemplate) {
     this.redisTemplate = redisTemplate;
     this.redisGetJob = redisGetJob;
     this.properties = properties;
-    this.hashService = hashService;
     this.cqlTemplate = cqlTemplate;
-    this.cassandraTemplate = cassandraTemplate;
-    this.appProperties = appProperties;
-    this.downsampleHashExistenceCache = downsampleHashExistenceCache;
-    this.timeslotExistenceCache = timeslotExistenceCache;
   }
 
   public Flux<String> claimJob(Integer partition, String group) {
@@ -119,53 +91,6 @@ public class DownsampleTrackingService {
     return Long.parseLong(timeslot) < nowSeconds - partitionWidth;
   }
 
-  public Publisher<?> track(String tenant, String seriesSetHash, Instant timestamp) {
-    if (!properties.isTrackingEnabled()) {
-      log.error("Partitions are not configured");
-      return Mono.empty();
-    }
-    final HashCode hashCode = hashService.getHashCode(tenant, seriesSetHash);
-    final int partition = hashService.getPartition(hashCode, properties.getPartitions());
-    final String setHash = encodeSetHash(tenant, seriesSetHash);
-    return saveDownsampling(partition, setHash).then(saveTimeslots(partition, timestamp));
-  }
-
-  private Mono<?> saveDownsampling(Integer partition, String setHash) {
-    final CompletableFuture<Boolean> result = downsampleHashExistenceCache.get(
-        new DownsampleSetCacheKey(partition, setHash),
-        (key, executor) ->
-            this.cassandraTemplate.insert(new Downsampling(partition, setHash))
-                .name("saveDownsampling")
-                .metrics()
-                .retryWhen(appProperties.getRetryInsertDownsampled().build())
-            .flatMap(s -> Mono.just(true))
-            .toFuture()
-    );
-    return Mono.fromFuture(result);
-  }
-
-  private Mono<?> saveTimeslots(int partition, Instant timestamp) {
-    return Flux.fromIterable(DateTimeUtils.getPartitionWidths(properties.getGranularities())).flatMap(
-        group -> {
-          final Instant normalizedTimeSlot = timestamp.with(new TemporalNormalizer(Duration.parse(group)));
-          return saveTimeslot(partition, group, normalizedTimeSlot.getEpochSecond());
-        }
-    ).then(Mono.empty());
-  }
-
-  private Mono<?> saveTimeslot(Integer partition, String group, Long timeslot) {
-    final CompletableFuture<Boolean> result = timeslotExistenceCache.get(
-        new TimeslotCacheKey(partition, group, timeslot),
-        (key, executor) -> {
-          log.trace("Saving timeslot {} {} {}", partition, group, timeslot);
-          return redisTemplate.opsForSet().add(encodeTimeslotKey(partition, group), timeslot.toString())
-              .flatMap(s -> Mono.just(true))
-              .toFuture();
-        }
-    );
-    return Mono.fromFuture(result);
-  }
-
   public Mono<?> deleteTimeslot(Integer partition, String group, Long timeslot) {
     return redisTemplate.opsForSet().remove(encodeTimeslotKey(partition, group), timeslot.toString())
         .flatMap(result -> {
@@ -180,10 +105,6 @@ public class DownsampleTrackingService {
     log.trace("getDownsampleSets {} {}", timeslot, partition);
     return this.cqlTemplate.queryForFlux(GET_HASHES_TO_DOWNSAMPLE, String.class, partition)
         .map(setHash -> buildPending(timeslot, setHash));
-  }
-
-  private static String encodeSetHash(String tenant, String setHash) {
-    return String.format("%s|%s", tenant, setHash);
   }
 
   private static String encodeTimeslotKey(int partition, String group) {
