+++ src/test/java/com/rackspace/ceres/app/services/DataWriteServiceTest.java
@@ -94,7 +94,7 @@ class DataWriteServiceTest {
   MetadataService metadataService;
 
   @MockBean
-  DownsampleTrackingService downsampleTrackingService;
+  IngestTrackingService ingestTrackingService;
 
   @Autowired
   SeriesSetService seriesSetService;
@@ -140,7 +140,7 @@ class DataWriteServiceTest {
           .thenReturn(Mono.empty());
       when(metadataService.storeMetadata(any(), any(), any(), any())).thenReturn(Mono.empty());
 
-      when(downsampleTrackingService.track(any(), anyString(), any()))
+      when(ingestTrackingService.track(any(), anyString(), any()))
           .thenReturn(Mono.empty());
 
       Instant metricTime = validMetricTime();
@@ -159,13 +159,13 @@ class DataWriteServiceTest {
       assertViaQuery(tenantId, metricQueryTime, seriesSetHash, metric);
 
       verify(metadataService).storeMetadata(tenantId, seriesSetHash, metric.getMetric(), metric.getTags());
-      verify(downsampleTrackingService).track(tenantId, seriesSetHash, metric.getTimestamp());
+      verify(ingestTrackingService).track(tenantId, seriesSetHash, metric.getTimestamp());
       verify(metadataService).updateMetricGroupAddMetricName(
           tenantId, metricGroup, metric.getMetric(), metric.getTimestamp().toString());
       verify(metadataService).updateDeviceAddMetricName(
           tenantId, resource, metric.getMetric(), metric.getTimestamp().toString());
 
-      verifyNoMoreInteractions(metadataService, downsampleTrackingService);
+      verifyNoMoreInteractions(metadataService, ingestTrackingService);
     }
 
     @Test
@@ -196,7 +196,7 @@ class DataWriteServiceTest {
       when(metadataService.storeMetadata(any(), any(), any(), any()))
           .thenReturn(Mono.empty());
 
-      when(downsampleTrackingService.track(any(), anyString(), any()))
+      when(ingestTrackingService.track(any(), anyString(), any()))
           .thenReturn(Mono.empty());
 
       Instant metricTime = validMetricTime();
@@ -233,10 +233,10 @@ class DataWriteServiceTest {
           tenant1, resource, metric1.getMetric(), metric1.getTimestamp().toString());
       verify(metadataService).updateDeviceAddMetricName(
           tenant2, resource, metric2.getMetric(), metric2.getTimestamp().toString());
-      verify(downsampleTrackingService).track(tenant1, seriesSetHash1, metric1.getTimestamp());
-      verify(downsampleTrackingService).track(tenant2, seriesSetHash2, metric2.getTimestamp());
+      verify(ingestTrackingService).track(tenant1, seriesSetHash1, metric1.getTimestamp());
+      verify(ingestTrackingService).track(tenant2, seriesSetHash2, metric2.getTimestamp());
 
-      verifyNoMoreInteractions(metadataService, downsampleTrackingService);
+      verifyNoMoreInteractions(metadataService, ingestTrackingService);
     }
 
     @Test
